+++ src/support/glyphorder.h
@@ -12,8 +12,8 @@ typedef struct {
 } glyph_order_entry;
 typedef glyph_order_entry *glyph_order_hash;
 
-int try_name_glyph(glyph_order_hash *glyph_order, uint16_t _id, sds name);
-void lookup_name(glyph_order_hash *glyph_order, uint16_t _gid, sds *field);
+int try_name_glyph(glyph_order_hash *glyph_order, int _id, sds name);
+void lookup_name(glyph_order_hash *glyph_order, int _gid, sds *field);
 void delete_glyph_order_map(glyph_order_hash *map);
 
 glyph_order_hash *caryll_glyphorder_from_json(json_value *root, caryll_dump_options *dumpopts);
